Correct the colour of the inversed tag prefix.
The element `o-teaser__tag-suffix` has never been coloured correctly
for teaser variants which changes the teaser background colour,
like the opinion teaser. We never noticed because `o-teaser__duration`
was mostly used (but deprecated in favour of a more generic
`o-teaser__tag-suffix`). In the last major we removed
`o-teaser__duration` but:

- a: didn't notice `o-teaser__tag-suffix` had an incorrect colour
in some cases
- b: missed a `o-teaser__duration` css selector

Fixes: https://github.com/Financial-Times/o-teaser/issues/149